Change points `tags` type to `std::uint32_t`
---------

+++ include/CLUEstering/data_structures/PointsDevice.hpp
@@ -41,7 +41,7 @@ namespace clue {
     device_buffer<TDev, std::byte[]> m_buffer;
     std::optional<device_buffer<TDev, value_type[]>> m_uncertainty_buffer;
     std::array<std::optional<device_buffer<TDev, value_type[]>>, Ndim> m_sigma_buffers;
-    std::optional<device_buffer<TDev, std::size_t[]>> m_tags_buffer;
+    std::optional<device_buffer<TDev, std::uint32_t[]>> m_tags_buffer;
     PointsView<Ndim, element_type> m_view;
     std::optional<std::size_t> m_nclusters;
     std::int32_t m_size;
@@ -203,7 +203,7 @@ namespace clue {
     /// @note Used as a deterministic tie-breaker in place of the raw array index, for cases where
     /// the order of the points is not guaranteed to be stable (e.g. when produced by an upstream
     /// parallel/concurrent stage)
-    void set_tags(std::span<std::size_t> tags);
+    void set_tags(std::span<std::uint32_t> tags);
 
   private:
     inline static constexpr std::size_t Ndim_ = Ndim;
